feat: Add MongoStore for session storage
This commit adds the `MongoStore` dependency and configures it to store session data in MongoDB. The `MongoStore` is used as the store option in the `express-session` middleware. The `mongoUrl` and `collectionName` options are set to connect to the MongoDB database and specify the collection name for storing session data. Additionally, the `maxAge` option is set to 7 days for the session cookie.