feat: Add route for retrieving session by ID
This commit adds a new route `/sessions/:id` to retrieve a session by its ID. It uses the `Session` model to query the database and populate the associated `user` field. If the session is not found, a 404 status is returned. If there is an error during the retrieval process, a 500 status is returned.