feat: Add route for adding users to collaborative documents
The code changes in this commit add a new route `/collab_doc/addUser/:docId` to the `collab_doc.ts` file. This route allows users to add themselves as collaborators to a specific document. It checks if the document exists and if the user is already a collaborator. If the document is already in the CollaboratedDoc collection, the user is added to the existing collaboration. Otherwise, a new collaboration is created. This enhances the functionality of the application by enabling collaborative document editing.