feat: Add personal field to User model
The code changes in this commit add a new field `personal` to the `User` model in `User.model.ts`. This field is of type `Schema.ObjectId` and allows users to store a personal reference. This change enhances the functionality of the model by providing a way for users to associate a personal object with their account.