feat: Add personal document for new users
The code changes in this commit add a personal document for new users in the `googleAuth.ts` file. This document is created when a new user signs up using Google authentication. It allows users to have a personal document associated with their account. This enhances the functionality of the application by providing a way for users to store personal information or notes.